Timing out pilot if it has reached maxtime
